creating and cleaning files
