+++ teleop2servo/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+* [PR-01](https://github.com/AGH-CEAI/teleop2servo/pull/1) - Added joint conterol from the keyboard
+
 ### Changed
 ### Deprecated
 ### Removed
